lines 12000-16000
