+++ src/compiler/checker/mod.rs
@@ -6,8 +6,10 @@ mod lines_2000_4000;
 pub use lines_2000_4000::*;
 mod lines_4000_8000;
 pub use lines_4000_8000::*;
-mod lines_8000_16000;
-pub use lines_8000_16000::*;
+mod lines_8000_12000;
+pub use lines_8000_12000::*;
+mod lines_12000_16000;
+pub use lines_12000_16000::*;
 mod lines_16000_18000;
 pub use lines_16000_18000::*;
 mod lines_18000_20000;
